map updates
* clean of of delayed
* fix issue with scrolling when mouse cursor was at the top of the list
